% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_expfactory_experiment}
\alias{process_expfactory_experiment}
\title{Process expfactory experiment JSON data.}
\usage{
process_expfactory_experiment(path, local = FALSE)
}
\arguments{
\item{path}{Path to data file}

\item{local}{TRUE means locally saved JSON (default=FALSE)}
}
\value{
Data frame (long format)
}
\description{
Most tasks in \href{https://expfactory.github.io/experiments/}{The Experiment Factory library} save their data as JSON.
This function parses JSON data generated by a task. Note that The Experiment Factory Docker container
also JSON encodes task data, so if the task encodes its data as JSON, use \code{local=FALSE} (the default)
to double-decode your data. Use \code{local=TRUE} for single-encoded JSON e.g. when saving task data locally.
}
\keyword{expfactory}
