% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{reverse_code_survey}
\alias{reverse_code_survey}
\title{Reverse code expfactory survey values}
\usage{
reverse_code_survey(df, rows, max)
}
\arguments{
\item{df}{data frame containing (long) data for one survey}

\item{rows}{integer vector containing row numbers of scores to be reverse coded}

\item{max}{integer value of highest possible score}
}
\value{
data frame
}
\description{
Reverse code expfactory survey values.
}
