% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_expfactory_experiment}
\alias{process_expfactory_experiment}
\title{Process expfactory experiment JSON data.}
\usage{
process_expfactory_experiment(path)
}
\arguments{
\item{path}{Path to data file}
}
\value{
Data frame (long format)
}
\description{
Process expfactory experiment JSON data.
}
\keyword{expfactory}
